requirements.txt
azure-functions
azure-identity
azure-mgmt-logic
azure-mgmt-network
azure-mgmt-resourcegraph